losowanie liczby calkowitej

+++ src/Main.java
@@ -3,7 +3,7 @@ public class Main {
         System.out.println("Hello world!");
         System.out.println("Witaj ponownie");
 
-        double wylosowana = Math.random()*10+1;
+        int wylosowana = (int)(Math.random()*10+1);
         System.out.println(wylosowana);
     }
 }
